Delete cleanup functions
Other than it being perceived as a good coding
convention, I haven't found evidence for my
purpose as to why I need to clean up modules'
resources before the program ends. I expect the
OS to do that for me.

I set up the source code where each file
represents a module, and for now these modules
persist for the lifetime of the program. So
when the program teminates, the the modules will
be cleaned up by the OS.