Fix delta time and move it to game loop
In enemies.cc, I was wrongly saving one variable
as a normal int instead of a 64 bit int, which
could cause issues if game runs for long enough.

enemies.cc and player.cc were calculating the
time delta on their own, but I am now opting
to avoid this redundancy, since the work
in progress bullets module will need that as well.