Fix bullets
Every bullet had id 0, and they were doing too
much damage (because they were not being
despawned immediately after making a hit).

Interesting to understand that despite the bullets
were all having the same ID, they were still being
saved in the memory pool and rendered correctly.
They were not being overwritten.

Looks like unordered_map allows multiple elements
with the same key. Looking into it, I learned
that unordered_map does not rely on key hash
do determine distinct elements because of
possible collisions, so it makes an additional
comparison with the structs when the hash is the
same. ¹

I also made some changes in the logic to avoid
a bullet hitting multiple enemies at the same
instant. For this to happen, enemies need to
overlap, which seldom occurs because they bounce
with each other when they collide, but the other
benefit is avoidining the logic to waste CPU
cycles doing unnecessary checks.

¹ https://stackoverflow.com/questions/45142799