Fix difficulty scaling
The difficulty was getting too high too fast.
While I was respecing the difficulty forumula from
the original JS counterpart ¹, I wasn't accounting
that the variable in the formula is the score
displayed on screen *divided by 100*.

In the original JS version, the score saved in
memory is just the number of enemies killed ²,
and on the UI I just added 2 zeroes to it.
In this version, I am using the score displayed on
screen as the actual score, so I need to divide it
by 100 to compute the difficulty correctly.

¹ https://git.nook.of.nicolaf.io/joy/smoc/-/blob/c429d496/src/core.js#L392
² https://git.nook.of.nicolaf.io/joy/smoc/-/blob/c429d496/src/core.js#L342