Reset enemy speed when bouncing over the edges
The buff that applies a push to an enemy upon
bullet impact is meant for the player to use
it to push enemies away when shooting at them.
When enough push is applied, the enemy's speed
is increased.

I programmed the game logic to have the enemy
redirect themselves straight to the player's
location when colliding with the window edge. The
undintended effect of this buff is that now
the pushed enemy goes straight to the player
at fast speed, so the buff is more harmful
than useful, especially when there are a lot
of bullets in the map applying a lot of push.

I'm fixing this problem by resetting the enemy's
speed back to normal when colliding at the
window edge.