Move fire rate logic to bullets module
Will be useful when implementing entities
other than the main player that can fire bullets
in the same manner as the player does.

Introducing the concept of a "bullet source" in
order to accomplish this change of structure.

Bullet sources can be instantiated and deleted in
the map. A bullet source has a position and
direction. Bullets will spawn from a bullet source
continuously in a predefined rate of fire,
adopting the source's direction.